+++ nuxt/index.js
@@ -1,57 +1,39 @@
-/*
-    Nuxt.js module for bootstrap-vue
-
-    Usage:
-        - Install both Bootstrap and bootstrap-vue packages
-        - Add this into your nuxt.config.js file:
-            {
-                modules: [
-                    [
-                      'bootstrap-vue/nuxt',
-                      {
-                        bootstrapCss: true,   // default is true
-                        bootstrapVueCss: true // default is true
-                      }
-                    ]
-                ]
-            }
-        - By default both bootstrap.css and boootstrap-vue.css are included. Set either
-          option to false to not include the compiled CSS file(s)
-        - The option 'css' has been deprecated in favour of 'bootstrapCss'
-*/
-
 const { resolve } = require('path')
 
-const PluginDefaluts = {
-  boostrapCss: true,
-  boostrapVueCss: true
+function pickFirst(...args) {
+  for (const arg of args) {
+    if (arg !== undefined) {
+      return arg
+    }
+  }
 }
 
-// Main module entry point
 module.exports = function nuxtBootstrapVue(moduleOptions = {}) {
   // Merge moduleOptions with default
-  const options = Object.assign({}, PluginDefaluts, moduleOptions)
-
-  // Deprecate css option in favour of boostrapCss option
-  if (options.css === false) {
-    options.bootstrapCss = false
-    delete options.css
-    try {
-      console.warn('BootstrapVue Plugin: option "css" deprecated, use "bootstrapCss" instead')
-    } catch (e) {}
+  const options = {
+    ...this.options.bootstrapVue,
+    ...moduleOptions
   }
 
-  // Conditionally require Bootstrap css if not explicitly disabled
-  if (options.bootstrapCss === true) {
-    // Add Bootstrap CSS to first in head
-    this.options.css.unshift('bootstrap/dist/css/bootstrap.min.css')
+  const bootstrapVueCSS = pickFirst(options.bootstrapVueCSS, options.bootstrapVueCss, options.bvCSS)
+  if (bootstrapVueCSS) {
+    // Add BootstrapVue CSS
+    this.options.css.unshift('bootstrap-vue/dist/bootstrap-vue.css')
   }
+
+  const bootstrapCSS = pickFirst(options.bootstrapCSS, options.bootstrapCss, options.css)
+  if (bootstrapCSS) {
+    // Add Bootstrap CSS
+    this.options.css.unshift('bootstrap/dist/css/bootstrap.css')
+  }
+
+  // Transpile src
+  this.options.build.transpile.push('bootstrap-vue/src')
 
   // Register plugin, pasing options to plugin template
   this.addPlugin({
     src: resolve(__dirname, 'plugin.template.js'),
-    fileName: 'bootstrap-vue.js',
-    options
+    fileName: 'bootstrap-vue.js'
   })
 }