+++ src/components/breadcrumb/breadcrumb-link.js
@@ -1,10 +1,10 @@
 import { mergeData } from 'vue-functional-data-merge'
 import pluckProps from '../../utils/pluck-props'
-import { assign } from '../../utils/object'
 import BLink, { propsFactory as linkPropsFactory } from '../link/link'
 import { htmlOrText } from '../../utils/html'
 
-export const props = assign(linkPropsFactory(), {
+export const props = {
+  ...linkPropsFactory(),
   text: {
     type: String,
     default: null
@@ -13,19 +13,11 @@ export const props = assign(linkPropsFactory(), {
     type: String,
     default: null
   },
-  active: {
-    type: Boolean,
-    default: false
-  },
-  href: {
-    type: String,
-    default: '#'
-  },
   ariaCurrent: {
     type: String,
     default: 'location'
   }
-})
+}
 
 // @vue/component
 export default {
@@ -38,8 +30,6 @@ export default {
     let componentData = { props: pluckProps(props, suppliedProps) }
     if (suppliedProps.active) {
       componentData.attrs = { 'aria-current': suppliedProps.ariaCurrent }
-    } else {
-      componentData.attrs = { href: suppliedProps.href }
     }
 
     if (!children) {